Add function definition for manifest_decode
