+++ manifest/provider/getproviderschema.go
@@ -7,7 +7,6 @@ import (
 	"context"
 
 	"github.com/hashicorp/terraform-plugin-go/tfprotov5"
-	"github.com/hashicorp/terraform-plugin-go/tftypes"
 )
 
 // GetProviderSchema function
@@ -15,19 +14,12 @@ func (s *RawProviderServer) GetProviderSchema(ctx context.Context, req *tfprotov
 	cfgSchema := GetProviderConfigSchema()
 	resSchema := GetProviderResourceSchema()
 	dsSchema := GetProviderDataSourceSchema()
+	funcSchema := GetFunctionsSchema()
 
 	return &tfprotov5.GetProviderSchemaResponse{
 		Provider:          cfgSchema,
 		ResourceSchemas:   resSchema,
 		DataSourceSchemas: dsSchema,
-		Functions: map[string]*tfprotov5.Function{
-			"hello_world2": {
-				Return: &tfprotov5.FunctionReturn{
-					Type: tftypes.String,
-				},
-				Summary:     "hello_world2 test",
-				Description: "hello_world2 test",
-			},
-		},
+		Functions:         funcSchema,
 	}, nil
 }
